Remove some files and start the BDTopo dep var for LCZ
