+++ classification/Lcz_BDTopo_v2_OSM/allCitiesBDTOPO_V2.csv
@@ -1,0 +1,1 @@
+"75101","75102","75103","75104","75105","75106","75107","75108","75109","75110","75111","75112","75113","75114","75115","75116","75117","75118","75119","75120","95480","77355","77356","91482","77357","77358","78484","94056","95487","77359","77360","93059","95488","77361","95489","78490","91495","77368","78497","77369","78498","77370","77371","77372","77373","77374","78499","95500","78501","78503","77376","95504","77377","77378","77379","78506","78505","91507","91508","95509","95510","77380","91511","92062","77381","91514","77382","78516","78517","77383","77384","77385","77386","77387","78518","77388","91519","78520","91521","78522","78524","91525","91526","77390","95527","78528","93063","95529","78530","93064","78531","77391","77392","77393","77394","92063","77395","94065","77396","77397","77398","91533","91534","95535","78536","77399","78537","91538","77400","77402","77403","95539","91540","95541","92064","95543","91544","78545","91546","77405","93066","77406","77401","77404","91549","78569","91547","77407","77408","78548","78550","78551","77409","77410","91552","91553","77411","77412","77413","95554","95555","91556","78557","77414","78558","78559","91560","77415","77416","78561","77417","78562","95563","77418","77419","94067","77420","77421","78564","78565","77423","77424","95566","77425","78567","94068","94069","91568","77426","77427","91570","78571","93070","77428","95572","77429","77430","91573","77431","95574","77432","78575","78576","91577","77433","77434","77435","77436","77437","91578","77438","91579","95580","91581","77439","77440","77441","77442","77443","77444","95582","94070","95584","95585","78586","91587","78588","77445","91589","77446","92071","77447","78590","78591","77448","95592","91593","77449","77450","95594","93071","92072","77451","77452","77453","77454","77455","78597","77456","95598","91599","91600","77457","78601","77458","77459","91602","93072","94071","92073","95604","78605","77460","95607","78609","95610","77461","95611","94073","77462","78615","78616","77463","77464","77465","77466","91617","78618","77468","91619","77469","77470","77471","78620","78621","93073","77473","78624","77474","77475","77476","77477","95625","77478","77479","77480","94074","95627","95628","91629","92075","77481","91631","77482","77483","77484","92076","95633","77486","91634","91635","93074","95637","77487","77489","78638","91639","78640","95641","77490","77492","77493","78642","78643","77494","91645","78646","78647","91648","91649","77495","95651","95652","78653","91654","78655","95656","77496","77498","91657","95658","91659","95660","77500","91661","77501","91662","94075","92077","94076","91666","77504","77505","77506","91667","93077","77507","92078","77508","94077","77509","77510","94078","77511","91671","77512","78672","77513","77514","93078","78674","95675","77515","95676","77516","93079","78677","77517","95678","77518","91679","95680","78681","95682","78683","77519","77520","94079","77521","91685","77522","77523","77524","77525","94080","77526","78686","91687","94081","77527","77528","78688","77529","77530","77531","77532","77533","91689","95690","77534","91691"
