+++ classification/Lcz_IAUIDF_OSM/configFile.json
@@ -1,1 +1,1 @@
-{"description":"Apply the OSM workflow","geoclimatedb":{"path":"/tmp/geoclimate_db;AUTO_SERVER=TRUE","delete":true},"input":{"osm":[]},"output":{"folder":"/home/decide/Data/URBIO/Donnees_brutes/LCZ/TrainingDataSets/Indicators/OSM"},"parameters":{"distance":1000,"indicatorUse":"URBAN_TYPOLOGY","svfSimplified":false,"prefixName":"","mapOfWeights":{"sky_view_factor":1,"aspect_ratio":1,"building_surface_fraction":1,"impervious_surface_fraction":1,"pervious_surface_fraction":1,"height_of_roughness_elements":1,"terrain_roughness_class":1},"hLevMin":3,"hLevMax":15,"hThresholdLev2":10}}
+{"description":"Apply the OSM workflow","geoclimatedb":{"path":"/tmp/geoclimate_db;AUTO_SERVER=TRUE","delete":true},"input":{"osm":["Gri\u00e8ges"]},"output":{"folder":"/home/decide/Data/URBIO/Donnees_brutes/LCZ/TrainingDataSets/Indicators/OSM"},"parameters":{"distance":1000,"indicatorUse":["URBAN_TYPOLOGY"],"svfSimplified":false,"prefixName":"","mapOfWeights":{"sky_view_factor":1,"aspect_ratio":1,"building_surface_fraction":1,"impervious_surface_fraction":1,"pervious_surface_fraction":1,"height_of_roughness_elements":1,"terrain_roughness_class":1},"hLevMin":3,"hLevMax":15,"hThresholdLev2":10}}
