+++ classification/calculateIndependentVariables.groovy
@@ -13,6 +13,10 @@ import org.orbisgis.orbisdata.datamanager.jdbc.h2gis.H2GIS
 import org.orbisgis.orbisdata.datamanager.jdbc.*
 import groovy.json.JsonOutput
 
+import groovy_generic_function
+
+ggf = new groovy_generic_function()
+
 // ##############################################################
 // ################### MAIN SCRIPT  #############################
 // ##############################################################
@@ -22,7 +26,7 @@ import groovy.json.JsonOutput
 /////////////////////////////////////////////////////////////////
 // Each argument is recovered in a variable understandable...
 def configFileWorkflowPath = this.args[0]
-def pathCitiesToTreat = this.args[1]
+def pathCitiesToTreat = configFileWorkflowPath + File.separator + this.args[1]
 def outputFolder = this.args[2]
 def data = this.args[3].toUpperCase()
 def indicatorUse = this.args[4].replaceAll(" ", "").split(",")
@@ -35,16 +39,16 @@ def dependentVariableColName = this.args[10]
 def geometryField = this.args[11]
 def sridDependentVarIndic = this.args[12]
 def pathToSaveTrainingDataSet = this.args[13]
-def lczValList = this.args[14].replaceAll(" ", "").split(",")
+def correspondenceTable = this.args[14].replaceAll(" ", "").split(",")
 def dependentVariable2ndColNameAndVal = this.args[15]
 Integer resetDataset = this.args[16]
 def scaleTrainingDataset = this.args[17]
 
 // Convert the list into a Map to have the key and corresponding value
-def lczValMap = [:]
-lczValList.each{
+def correspondenceValMap = [:]
+correspondenceTable.each{
 	def keyAndVal = it.split(":")
-	lczValMap[keyAndVal[0]] = keyAndVal[1]
+	correspondenceValMap[keyAndVal[0]] = keyAndVal[1]
 }
 
 // Define some table names
@@ -55,10 +59,10 @@ def dependentVarTableNameRaw = "dependentVarTableName_raw"
 // Start to operate processes //////////////////////////////////////
 ////////////////////////////////////////////////////////////////////
 // Prepare and launch the workflow
-executeWorkflow(configFileWorkflowPath, pathCitiesToTreat, outputFolder, data, indicatorUse, dbUrl, dbId, dbPassword, resetDataset)
+ggf.executeWorkflow(configFileWorkflowPath, pathCitiesToTreat, outputFolder, data, indicatorUse, dbUrl, dbId, dbPassword, resetDataset)
 
 // Open a database used for calculation and load the IAUIdF file (LCZ)
-H2GIS datasource = H2GIS.open("/tmp/classification${getUuid()};AUTO_SERVER=TRUE", "sa", "")
+H2GIS datasource = H2GIS.open("/tmp/classification${ggf.getUuid()};AUTO_SERVER=TRUE", "sa", "")
 datasource.load(dependentVariablePath, dependentVarTableNameRaw)
 
 // Set the Srid of the independent variables dataset according to the dataset origin
@@ -80,14 +84,14 @@ datasource.getSpatialTable(dependentVarTableName).the_geom.createIndex()
 datasource.getSpatialTable(dependentVarTableName)."$dependentVariableColName".createIndex()
 
 // Gather indicators from different scales (building, block, RSU) into a single table at the scale of the training dataset
-createTrainingDataset(datasource, pathCitiesToTreat, outputFolder, data, operationsToApply, dependentVarTableName, dependentVariableColName, pathToSaveTrainingDataSet, lczValMap, dependentVariable2ndColNameAndVal, resetDataset, scaleTrainingDataset)
+createTrainingDataset(datasource, pathCitiesToTreat, outputFolder, data, operationsToApply, dependentVarTableName, dependentVariableColName, pathToSaveTrainingDataSet, correspondenceValMap, dependentVariable2ndColNameAndVal, resetDataset, scaleTrainingDataset)
 
 // 
 
 // #####################################################################
 // ################### FUNCTIONS TO USE  ###############################
 // #####################################################################
-def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, String outputFolder, String data, String[] operationsToApply, String dependentVarTableName, String dependentVariableColName, String pathToSaveTrainingDataSet, def lczValMap, String dependentVariable2ndColNameAndVal, Integer resetDataset, String scaleTrainingDataset){
+def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, String outputFolder, String data, String[] operationsToApply, String dependentVarTableName, String dependentVariableColName, String pathToSaveTrainingDataSet, def correspondenceValMap, String dependentVariable2ndColNameAndVal, Integer resetDataset, String scaleTrainingDataset){
 	/**
 	* For each city, gather indicators from different scales (building, block, RSU) into a single table at RSU scale and then associate the dependent variable value for the corresponding location
 	* @param configFileWorkflowPath 	The path of the config file
@@ -203,10 +207,14 @@ def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, S
 				listScal1Rename = []
 				listScal2Rename = []
 				for (col in scale1Col2Rename){
-					listScal1Rename.add("a.$col AS build_$col")
+					if(col != "id_build" || col != "id_block" || col != "id_rsu"){
+						listScal1Rename.add("a.$col AS build_$col")
+					}
 				}
 				for (col in scale2Col2Rename){
-					listScal2Rename.add("b.$col AS block_$col")
+					if(col != "id_block" || col != "id_rsu"){
+						listScal2Rename.add("b.$col AS block_$col")
+					}
 				}
 
 				// Define generic name whatever be the 'scaleTrainingDataset'	
@@ -223,10 +231,14 @@ def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, S
 				listScal1Rename = []
 				listScal2Rename = []
 				for (col in scale1Col2Rename){
-					listScal1Rename.add("a.$col AS block_$col")
+					if(col != "id_block" || col != "id_rsu"){
+						listScal1Rename.add("a.$col AS block_$col")
+					}
 				}
 				for (col in scale2Col2Rename){
-					listScal2Rename.add("b.$col AS rsu_$col")
+					if(col != "id_rsu"){
+						listScal2Rename.add("b.$col AS rsu_$col")
+					}
 				}
 				// Define generic name whatever be the 'scaleTrainingDataset'	
 				finalScaleTableName = buildingIndicTempo
@@ -258,7 +270,7 @@ def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, S
 			def areaCalcQuery = "ST_AREA(ST_INTERSECTION(a.the_geom,b.the_geom))"
 			def casewhenQuery = ""
 			def sumQuery = ""
-			lczValMap.eachWithIndex{ind, val ->
+			correspondenceValMap.each{ind, val ->
 				casewhenQuery+="CASE WHEN a.$dependentVariableColName='${ind}' THEN ${areaCalcQuery} ELSE 0 END AS $prefixDependentVar${val}, "
 				sumQuery += "SUM($prefixDependentVar${val}) AS $prefixDependentVar${val}, "
 			}
@@ -293,7 +305,7 @@ def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, S
 			datasource.getTable(allFinalScaleIndic).id_rsu.createIndex()
 			datasource.execute """ DROP TABLE IF EXISTS $trainingData;
 									CREATE TABLE $trainingData
-											AS SELECT a.EXTREMUM_COL AS LCZ, a.UNIQUENESS_VALUE, b.*
+											AS SELECT a.EXTREMUM_COL AS $dependentVariableColName, a.UNIQUENESS_VALUE, b.*
 											FROM $resultsDistrib a 
 												RIGHT JOIN $allFinalScaleIndic b
 												ON a.id_rsu = b.id_rsu;
@@ -307,195 +319,5 @@ def createTrainingDataset(JdbcDataSource datasource, String pathCitiesToTreat, S
 			println("'$areaName' has been processed")
 		}
 	}
-}
-
-
-def executeWorkflow(String configFileWorkflowPath, String pathCitiesToTreat, String outputFolder, String data, String[] indicatorUse, String dbUrl, String dbId, String dbPassword, Integer resetDataset){
-	/**
-	* Prepare and launch the Workflow (OSM or BDTOPO_V2) configuration file
-	* @param configFileWorkflowPath 	The path of the config file
-	* @param pathCitiesToTreat		Path of the file where are stored the cities to process
-	* @param outputFolder 			Path where are saved all intermediate and output results
-	* @param data				The type of dataset used for the calculation of independent variables (OSM or BDTOPO_V2)
-	* @param indicatorUse 			A given batch of indicators will be calculated depending on what is (are) the indicator use(s)
-	* @param dbUrl				The URL of the database to download the data (in case the dataset is not OSM)
-	* @param dbId 				The username (or id) for the database access
-	* @param dbPassword			The password for the database access
-	* @return				None
-	*/
-	// The  results are saved in a specific folder depending on dataset type
-	outputFolder = outputFolder + data
-	
-	// Modify the input db parameters (if they have "default" as value, it means they are supposed to be "")
-	if(dbUrl=="default"){
-		dbUrl = ""
-	}
-	if(dbId=="default"){
-		dbId = ""
-	}
-	if(dbPassword=="default"){
-		dbPassword = ""
-	}
-
-	// Create the config file that will be used for the calculation of OSM Workflow
-	File dirFile = new File(configFileWorkflowPath)
-	dirFile.delete()
-	dirFile.mkdir()
-
-	// Load the file containing all cities that need to be done
-	def line
-	new File(pathCitiesToTreat).withReader('UTF-8') { reader ->
-	    line = reader.readLine().replace('"', '')
-	}
-	def allCities = line.split(",")
-
-	// Load cities already done
-	resultsGeoclimate=outputFolder
-	File[] citiesAlreadyDone = new File(resultsGeoclimate).listFiles()
-
-	// If 'resetDataset', remove cities already done from the list to be done
-	def listToBeDone = allCities
-	if(resetDataset != 0){
-		for (pathCity in citiesAlreadyDone){
-			city = pathCity.toString().split("/").last()[data.size()+1..-1]
-			listToBeDone-=city
-		}
-	}
-
-	if(listToBeDone.size()==0){
-		println "$data worflow: no new city need to be calculated"
-	}
-	else{
-		println "$data worflow: the following cities will be calculated:"
-		println Arrays.toString(listToBeDone)
-
-		// Create the config file
-		// Convert the 'indicatorUse' parameter to array of string
-
-		def worflow_parameters	
-		if(data=="OSM"){
-			worflow_parameters = [
-				    	"description" :"Apply the $data workflow",
-				    	"geoclimatedb" : 		[
-									"path" : "/tmp/geoclimate_db;AUTO_SERVER=TRUE",
-									"delete" :true
-						    			],
-				    	"input" : 			[
-									"osm" : listToBeDone
-									],
-				    	"output" :			[
-				     					"folder" : outputFolder
-									],
-					"parameters":
-								    	[
-									"distance" : 1000,
-									"indicatorUse": indicatorUse,
-									"svfSimplified": false,
-									"prefixName": "",
-									"mapOfWeights":
-											["sky_view_factor": 1,
-											    "aspect_ratio": 1,
-											    "building_surface_fraction": 1,
-											    "impervious_surface_fraction" : 1,
-											    "pervious_surface_fraction": 1,
-											    "height_of_roughness_elements": 1,
-											    "terrain_roughness_class": 1
-											],
-									"hLevMin": 3,
-									"hLevMax": 15,
-									"hThresholdLev2": 10
-						    			]
-					]
-		}
-		if(data=="BDTOPO_V2"){
-			worflow_parameters = [
-						"description" :"Apply the $data workflow",
-						"geoclimatedb" : 		[
-										"path" : "/tmp/geoclimate_db;AUTO_SERVER=TRUE",
-										"delete" :true
-							    			],
-						"input" : 			[
-										"database": [
-											    "user": dbId,
-											    "password": dbPassword,
-											    "url": dbUrl,
-											    "id_zones":	listToBeDone,
-											    "tables": 	[
-													"iris_ge":"ign_iris.iris_ge_2016",
-													"bati_indifferencie":"ign_bdtopo_2017.bati_indifferencie",
-													"bati_industriel":"ign_bdtopo_2017.bati_industriel",
-													"bati_remarquable":"ign_bdtopo_2017.bati_remarquable",
-														"route":"ign_bdtopo_2017.route",
-													"troncon_voie_ferree":"ign_bdtopo_2017.troncon_voie_ferree",
-														"surface_eau":"ign_bdtopo_2017.surface_eau",
-													"zone_vegetation":"ign_bdtopo_2017.zone_vegetation",
-														"terrain_sport":"ign_bdtopo_2017.terrain_sport",
-													"construction_surfacique":"ign_bdtopo_2017.construction_surfacique",
-														"surface_route":"ign_bdtopo_2017.surface_route",
-													"surface_activite":"ign_bdtopo_2017.surface_activite"
-													]
-												]
-										],
-						"output" :			[
-										"/home/decide/Data/URBIO/Donnees_brutes/LCZ/TrainingDataSets/Indicators/BDTOPO_V2"
-										],
-						"parameters":			[
-										"distance" : 1000,
-										"indicatorUse": indicatorUse,
-										"svfSimplified": false,
-										"prefixName": "",
-										"hLevMin": 3,
-										"hLevMax": 15,
-										"hThresholdLev2": 10
-										]
-						]
-		}
-		// Fill the workflow config parameter and return the path
-		def configFilePath = createOSMConfigFile(worflow_parameters, configFileWorkflowPath)
-
-		// Produce the indicators for the entire Ile de France and save results in files
-		produceIndicatorsForTest(configFilePath, data)
-	}
-}
-
-/**
-* Execute several cities where we have testsIProcess process = ProcessingChain.Workflow.BDTOPO_V2()
-* @param configFile 	The path of the config file
-* @param data		The type of dataset (OSM or BDTOPO_V2)
-* @return		None
-*/
-void produceIndicatorsForTest(String configFile, String data) {
-	if (data == "OSM"){	
-		def process = Geoclimate.OSM.workflow
-		process.execute(configurationFile: configFile)
-	}
-	else if (data == "BDTOPO_V2"){
-		def process = Geoclimate.BDTOPO_V2.workflow
-		process.execute(configurationFile: configFile)
-	}
-}
-
-/**
-* Simple function to create a large number to set random filename 
-* @return the number as String
-*/
-static def getUuid(){
-    UUID.randomUUID().toString().replaceAll("-", "_") }
-
-/**
-* Create a configuration file
-* @param 	osmParameters
-* @param 	directory
-* @return	the path of the config file name
-*/
-def createOSMConfigFile(def worflow_parameters, def directory){
-	def json = JsonOutput.toJson(worflow_parameters)
-	def configFilePath =  directory+File.separator+"configFile.json"
-	File configFile = new File(configFilePath)
-	if(configFile.exists()){
-	    configFile.delete()
-	} 
-	configFile.write(json)
-	return configFile.absolutePath
 }
